[TE] Support Sugon/TianLong SHCA IB
Add USE_SHCA build option for Sugon/TianLong SHCA NICs, which use
extended 17-bit LIDs instead of standard 16-bit IB LIDs. When enabled,
convert LIDs via shca_17b_types.h helpers and widen lid fields to
uint32_t across Transfer Engine and TENT RDMA paths.

Gate yalantinglibs IBV RPC (YLT_ENABLE_IBV) behind USE_SHCA=OFF so SHCA
builds skip ib_socket-based RDMA RPC and fall back to TCP with a warning.

Adjust dependencies.sh to detect SHCA headers and skip conflicting system
RDMA packages (libibverbs-dev on Debian, rdma-core-devel on RHEL).

+++ mooncake-transfer-engine/include/transport/rdma_transport/rdma_context.h
@@ -144,7 +144,7 @@ class RdmaContext {
     std::string nicPath() const;
 
    public:
-    uint16_t lid() const { return lid_; }
+    uint32_t lid() const { return lid_; }
 
     std::string gid() const;
 
@@ -209,7 +209,7 @@ class RdmaContext {
     ibv_comp_channel **comp_channel_ = nullptr;
 
     uint8_t port_ = 0;
-    uint16_t lid_ = 0;
+    uint32_t lid_ = 0;
     int gid_index_ = -1;
     int active_speed_ = -1;
     ibv_mtu active_mtu_;
